feat(category): update category repository

+++ src/Infrastructure/Data/Repositories/CategoryRepository.cs
@@ -1,7 +1,9 @@
-﻿using API_Manajemen_Barang.src.Application.Interfaces;
+﻿using API_Manajemen_Barang.src.Application.Exceptions;
+using API_Manajemen_Barang.src.Application.Interfaces;
 using Inventory_api.src.Core.Entities;
 using Inventory_api.src.Infrastructure.Data;
 using Microsoft.EntityFrameworkCore;
+using System.Threading.Tasks;
 
 namespace API_Manajemen_Barang.src.Infrastructure.Data.Repositories
 {
@@ -16,24 +18,35 @@ namespace API_Manajemen_Barang.src.Infrastructure.Data.Repositories
 
         public async Task<IEnumerable<Category>> GetAllCategoryAsync()
         {
-            return await _context.Categories.Include(i => i.Items).ToListAsync();
+            return await _context.Categories
+                .AsNoTracking()
+                .Include(i => i.Items)
+                .ToListAsync();
         }
 
         public async Task<Category?> GetByIdAsync(int id)
         {
-            return await _context.Categories.Include(i => i.Items).FirstOrDefaultAsync(c => c.CategoryId == id);
+            return await _context.Categories
+                .AsNoTracking()
+                .Include(i => i.Items)
+                .FirstOrDefaultAsync(c => c.CategoryId == id);
         }
 
         public async Task<Category?> GetByNameAsync(string name)
         {
-            return await _context.Categories.Include(i => i.Items)
+            return await _context.Categories
+                .AsNoTracking()
+                .Include(i => i.Items)
                 .Where(c => c.Name.ToLower().Contains(name.ToLower()))
                 .FirstOrDefaultAsync();
         }
 
         public async Task<bool> IsCategoryExist(string name)
         {
-            return await _context.Categories.AnyAsync(c => c.Name.ToLower() == name.ToLower());
+            return await _context.Categories
+                .AsNoTracking()
+                .Include(i => i.Items)
+                .AnyAsync(c => c.Name.ToLower() == name.ToLower());
         }
 
         public async Task AddAsync(Category category)
@@ -46,10 +59,19 @@ namespace API_Manajemen_Barang.src.Infrastructure.Data.Repositories
             _context.Categories.Update(category);
         }
 
-        public void DeleteAsync(int id)
+        public async Task DeleteAsync(int id)
         {
-            var category = new Category { CategoryId = id };
-            _context.Categories.Attach(category);
+            if (id <= 0)
+            {
+                throw new BadRequestException($"ID {id} Invalid. Please check and try again!");
+            }
+
+            var category = await _context.Categories.FindAsync(id);
+            if (category == null)
+            {
+                throw new NotFoundException($"Cattegory with ID {id} not found");
+            }
+
             _context.Categories.Remove(category);
         }
     }
